repo

+++ src/Core/HR.LeaveManagement.Application/Features/LeaveType/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,6 +1,7 @@
 using System;
 using AutoMapper;
 using HR.LeaveManagement.Application.Contracts.Persistance;
+using HR.LeaveManagement.Application.Exceptions;
 using MediatR;
 
 namespace HR.LeaveManagement.Application.Features.LeaveType.Commands.DeleteLeaveType;
@@ -19,11 +20,11 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
         //retrieve domaim entity object
         var leaveTypeToDelete = await _leaveTypeRepository.GetByIdAsync(request.Id);
 
-        //verify that record is exist
-        // if (leaveTypeToDelete == null)
-        // {
-
-        // }
+        //verify that record is not exist
+        if (leaveTypeToDelete == null)
+        {
+            throw new NotFoundException(nameof(LeaveType), request.Id);
+        }
 
         //remove from Database
         await _leaveTypeRepository.DeleteAsync(leaveTypeToDelete);
